Kevin development (#75)
* Committing to my own branch to merge with Development for Tyler's updates

* Updated Database manager

Added description update, insert, and get functions

* updated databasemananger

* Changed path to properties

Hard-coded path to properties for a good local connection

* Keeping changes to DatabaseManager

* Added the working DatabaseManager back into the package

* name -> dataName fix

* Committing to Brandon-Development to continue working from lab

* Toyed with DatabaseManager (then undid all my changes), tried managing the user table with UserManager

* Beginning Admin

* Unsure of Dan's progress, pushing ahead with admin functions and GUI

* More work on admin functionality

* More Admin progress

No functionality, add entry button not connected to a method. Beginning
work on Delete Data tab

* Admin work

* Updated styles for Dashboard GUI

* fixed the limits

* Group development (#48)

Group development

* changed the export tab to a export button that blends in with the other
tabs.

* Added double Y-Axis and default to Temperature of Air and Water

* Began revising so that the data parameter chosen can  is based on it's actual id

* Refactored DataReceiver a bit so that it maintains a single PARAMETER_MAP, which keeps track of DataParameters, which is an object that keeps track of the name, id, unit, etc., for a given parameter

* Updated .gitignore

Moving the GitHub folder to My Documents caused a small issue with
thumbs.db, this resolves that issue

* Started playing around with Highchartjs in order to eventually make the
graph zoomable with a double axis for 2 data points

* Added a way to get just raw data from the Data returned from DataReceiver for people who would prefer to use it over Observables

* Styled the Highcharts chart to match the styling of the rest of the page,
and and played around trying to get real data in there

* Updated the chart so that it pulls data from the request attributes, so Kevin can focus purely on front end and not backend.

* Took the Highchart graph that Louis was able to make to generate real data
and styled it to zoom, resize, and a couple other minor changes.

* Created body of admin AJAX/jQuery insertion template

Using the example I followed online, I typed up basic insertion and
deletion functions to test. Now I am translating it to our project,
changing variables as necessary. Testing with our JSON's comes next.

* Made it so there is a dedicated form for the graph and table tab
also made it so that whichever tab was submitted from is the tab that 
loads after submission

* Change in properties: WaterQualityServer now WaterQuality

I changed the name of the project per Dr. Jones' request. It is no
longer WaterQualityServer. The properties files need to be kept to keep
the change.

Also committing work on the javascript files to implement AJAX requests
on the dashboard.

* Limited X-Axis to display 5 at a time (dynamically based on amount of data).

* Began to create method to process AJAX query

* Correctly implemented Protocol, as well as fixing up graph

* Ready to commit current changes; finished adding protocol to both JavaScript AND Java

* Resolved an issue where I improperly merged into Group-Development, works now

* Just changed it back so that table still generates when  graph is made

* just fixing stuff

* L.J: Fixed some issues for Kevin

* added the capability for double axes. Made it display with the color of
the lines and dynamically places names of data based on whatever is checked

* AJAX generic function defined!

The generic AJAX function is defined in AJAX_magic.js

ControlServlet and dashboard are updated for testing passing data
between the two pages using the new function defined above.

descriptions.json is fixed to read the correct data, thanks to Dr.
Rier's keen eye.

* Dan-Development wk4 (#67)

* Admin Pages minor fixes

Minor changes on Dan's and began some changes on Brandon's

* Styling and Design

Made changes to the styling and design of Brandon's version of the admin
pages.  Also did some renaming of CSS classes.

* Fixed fluctations in graph

* Admin progress...

Beginning attempt to connect to AdminServlet from admin page...

* Fixed files for correct data entry

Debugged for over an hour with Brett's and Jake's help to get a
successful connection to Tyler's database methods, and into the table.
*phew* Testing code has not been taken out yet.

* Added documentation, as well as a specific post/get function to extend on Brandon's post_get.

* trying to make it so the dashboard container will be centered vertically

* Can now successfully obtain data from the server! Huge improvement, thanks to Brandon's AJAX function

* Updated the graph so that it updates dynamically: Currently the table is not properly generated, need to do that.

* Managed to implement date time for HighChart

* created a dashboard.js file for the javascript functions
of the dashboard. Also made some changes including making the graph 
checkboxes dynamically unclick using a queue if too many are clicked.
Also made it so that checkboxs no longer clear when you change tab.
So each tab maintains the state of its checkboxes.

* trying to figure out the table

* made so the buttons are styled again

* worked on getting the table to work. Almost works. Still need to fix an
issue however.

* The descriptions and table now update, but these are very temporary solutions to a long term problem

* Made it so the table tab generates data separately from the graph tab.
Also did some more work on the new fillTable() function. It successfuly
creates a table of data for certain value pairs but not all because of 
missing data.