Made Remove User and Modified Errors tab
Made Remove User and it is fully functioning.

Errors tab now displays all errors within a month to start out, and
updates the graph whenever a new time is selected instead of after
display button is pressed.

CSS is fixed so the data tables on the admin pages don't look like
garbage.